requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
langchain-aws
langchain-community 
langchain-core
awscli
boto3